requirements.txt
django==2.2.14
djangorestframework==3.11.0
psycopg2-binary==2.8.1
django-rest-swagger==2.2.0
Pillow==7.2.0
python-dotenv==0.10.1
